cart id生成用雪花，修改类型

+++ src/main/java/com/bupt/uta/entity/CartVo.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CartVo {
-    private int cartId;
+    private Long cartId;
     private Long customerId;
     private int productId;
     private int status;
